todo/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
asttokens==2.2.1
async-timeout==3.0.1
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
channels-redis==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
Cython==0.29.32
daphne==3.0.2
decorator==5.1.1
Django==4.1.4
django-celery-results==2.4.0
django-rest-knox==4.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
executing==1.2.0
greenlet==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.11
ipython==8.7.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
msgpack==1.0.4
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytesseract==0.3.10
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.5.2
